BROKEN - renamed card.js to .ts (for history)
